moved AtomTransmuter class
