--- MDANSE/Src/MDANSE/Framework/AtomTransmutation/__init__.py
@@ -1,1 +1,0 @@
-from .transmutation import AtomTransmuter
